updated changes in editFaq and deleteFaq api's

+++ services/admin/static_content.service.js
@@ -44,41 +44,39 @@ module.exports = {
         
      },
 
-     /*deleteFaq: async (params, user) => {
-        const faqData = await Faq.destroy({
-            where: {
-                //topic_id:Number(req.body.topic_id),
-                topic_id:Number(params.topic_id),
-                admin_id:String(user.id)
-            } 
-        })
-        const faqTopicsData = await FaqTopics.destroy({
-            where: {
-                id:Number(params.topic_id)
-            } 
-        })
-    ],*/
           
      deleteFaq: async(params,admin)=>{
-        const faqData = await Faq.destroy({
-            where: {
-                //topic_id:Number(req.body.topic_id),
-                topic_id:Number(params.topic_id),
-                admin_id:String(admin.id)
-            } 
+        let checkTopicId = await Faq.findOne({
+            where: { topic_id: params.topic_id}
         })
-        const faqTopicsData = await FaqTopics.destroy({
-            where: {
-                id:Number(params.topic_id)
-            } 
-        })
+        if (checkTopicId) {
+            const faqData = await Faq.destroy({
+                where: {
+                    //topic_id:Number(req.body.topic_id),
+                    topic_id:Number(params.topic_id),
+                    admin_id:String(admin.id)
+                } 
+            })
+            return true
+        } else {
+            throw new Error(constants.MESSAGES.invalid_id);
+        }
      },
 
      editFaq: async(params,admin,topicId)=>{
-        const faqData=await Faq.update({ question:params.question,answer:params.answer }, { where: {topic_id:Number(topicId),admin_id:String(admin.id)} });
+        let checkTopicId = await Faq.findOne({
+            where: { topic_id: topicId}
+        })
+        if (checkTopicId) {
+            const faqData=await Faq.update({ question:params.question,answer:params.answer }, { where: {topic_id:Number(topicId),admin_id:String(admin.id)} });
+            return true
+        } else {
+            throw new Error(constants.MESSAGES.invalid_id);
+        }
+       /* const faqData=await Faq.update({ question:params.question,answer:params.answer }, { where: {topic_id:Number(topicId),admin_id:String(admin.id)} });
         console.log(faqData)
       const faqTopicsData=await FaqTopics.update({ topic:params.topic }, { where: {id:Number(topicId)} });
-        console.log(faqTopicsData)
+        console.log(faqTopicsData)*/
      },
  
      getFaqs: async () => {
